use SLIP10 for Ed25519 key derivation frontend
