+++ sdk/typescript/package.json
@@ -86,9 +86,12 @@
     "@mysten/bcs": "workspace:*",
     "@noble/hashes": "^1.1.2",
     "@noble/secp256k1": "^1.6.3",
+    "@scure/bip32": "^1.1.0",
+    "bip39-light": "^1.0.7",
     "bn.js": "^5.2.0",
     "buffer": "^6.0.3",
     "cross-fetch": "^3.1.5",
+    "ed25519-hd-key": "^1.3.0",
     "jayson": "^3.6.6",
     "js-sha3": "^0.8.0",
     "lossless-json": "^1.0.5",
