+++ sdk/typescript/src/types/bip39-light.d.ts
